Organizing components into separate files
